requirements.txt

pandas>=1.3
scikit-learn>=1.0
matplotlib>=3.4
seaborn>=0.11
google-cloud-aiplatform>=1.26
kfp>=2.0.0
joblib>=1.1
